third_party/WebKit/LayoutTests/editing/spelling/spellcheck-attribute-settings-default-expected.txt

This tests that the spellcheck default value is as specified in internal blink settings. This allows embedders to specify spellchecking behavior when the spellcheck attribute is not explicitly set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


id=test1_1 type=text spellcheck=true enabled_by_default=true
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_2 type=text spellcheck=true enabled_by_default=true
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_3 type=text spellcheck=false enabled_by_default=true
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_4 type=text spellcheck=true enabled_by_default=true
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_5 type=text spellcheck=true enabled_by_default=true
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_1 type=text spellcheck=false enabled_by_default=false
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_2 type=text spellcheck=true enabled_by_default=false
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_3 type=text spellcheck=false enabled_by_default=false
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_4 type=text spellcheck=false enabled_by_default=false
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
id=test1_5 type=text spellcheck=true enabled_by_default=false
PASS internals.hasSpellingMarker(document, 0, 2) became shouldBeMarked
PASS successfullyParsed is true

TEST COMPLETE
